+++ src/main/scala/sangria/validation/rules/KnownDirectives.scala
@@ -52,14 +52,21 @@ object KnownDirectives {
     case _: ast.InlineFragment ⇒ Some(DirectiveLocation.InlineFragment → "inline fragment")
 
     case _: ast.SchemaDefinition ⇒ Some(DirectiveLocation.Schema → "schema definition")
+    case _: ast.SchemaExtensionDefinition ⇒ Some(DirectiveLocation.Schema → "schema extension definition")
     case _: ast.ScalarTypeDefinition ⇒ Some(DirectiveLocation.Scalar → "scalar type definition")
+    case _: ast.ScalarTypeExtensionDefinition ⇒ Some(DirectiveLocation.Scalar → "scalar type extension definition")
     case _: ast.ObjectTypeDefinition ⇒ Some(DirectiveLocation.Object → "object type definition")
+    case _: ast.ObjectTypeExtensionDefinition ⇒ Some(DirectiveLocation.Object → "object type extension definition")
     case _: ast.FieldDefinition ⇒ Some(DirectiveLocation.FieldDefinition → "field definition")
     case _: ast.InterfaceTypeDefinition ⇒ Some(DirectiveLocation.Interface → "interface definition")
+    case _: ast.InterfaceTypeExtensionDefinition ⇒ Some(DirectiveLocation.Interface → "interface extension definition")
     case _: ast.UnionTypeDefinition ⇒ Some(DirectiveLocation.Union → "union definition")
+    case _: ast.UnionTypeExtensionDefinition ⇒ Some(DirectiveLocation.Union → "union extension definition")
     case _: ast.EnumTypeDefinition ⇒ Some(DirectiveLocation.Enum → "enum definition")
+    case _: ast.EnumTypeExtensionDefinition ⇒ Some(DirectiveLocation.Enum → "enum extension definition")
     case _: ast.EnumValueDefinition ⇒ Some(DirectiveLocation.EnumValue → "enum value definition")
     case _: ast.InputObjectTypeDefinition ⇒ Some(DirectiveLocation.InputObject → "input object type definition")
+    case _: ast.InputObjectTypeExtensionDefinition ⇒ Some(DirectiveLocation.InputObject → "input object type extension definition")
     case _: ast.InputValueDefinition ⇒
       parent match {
         case _: ast.InputObjectTypeDefinition ⇒ Some(DirectiveLocation.InputFieldDefinition → "input field definition")